Updates the FractionalQuantityRegressionAlgorithm statistics
This regression test presents itself occassionaly as a flickering test.
It seems the statistics were changed during such a flicker incorrectly.
By looking at the design of the algorith, one would expect it to trade
on consecutive days, which these statistcs represent. With the recent
filtering changes the flicker rate seems to be waning.

Original Analysis (might have some truths, left for posterity):
The issue stems from the bitcoin reader returning default instances instead of
null. These default instances get piped all the way through to the algorithm.
While this is happening, the trusty ol' benchmark security, SPY, is keeping
real time. The race is whether or not the SPY subscription can pass the first
real date in the bitcoin data. If this happens, multiple bitcoin dates would
all be dequeued at once since their in the past.

This change solves the race condition by filtering out invalid data points as
early as possible in the pipeline.